--- src/routines/Routine.ts
@@ -1,12 +1,0 @@
-export abstract class Routine {
-  protected interval: number;
-  protected abstract routineFunction(): void | Promise<void>;
-
-  constructor(interval?: number) {
-    this.interval = interval ?? 60;
-  }
-
-  start() {
-    setInterval(this.routineFunction, this.interval * 1000);
-  }
-}
